extract task class into task.cxx
