+++ src/core/promise.cxx
@@ -10,6 +10,7 @@ import std;
 import libfork.utils;
 import :concepts;
 import :frame;
+import :task;
 import :thread_locals;
 import :ops;
 import :handles;
@@ -23,44 +24,6 @@ using coro = std::coroutine_handle<T>;
 
 template <worker_context Context>
 using frame_t = frame_type<checkpoint_t<Context>>;
-
-// =============== Forward-decl =============== //
-
-// TODO: public private split here and elsewhere
-export template <returnable T, worker_context Context>
-struct promise_type;
-
-// =============== Task =============== //
-
-/**
- * @brief The return type for libfork's async functions/coroutines.
- *
- * This predominantly exists to disambiguate `libfork`s coroutines from other
- * coroutines and specify `T` the async function's return type which is
- * required to be `void` or a `std::movable` type.
- *
- * \rst
- *
- * .. note::
- *
- *    No consumer of this library should ever touch an instance of this type,
- *    it is used for specifying the return type of an `async` function only.
- *
- * \endrst
- */
-export template <returnable T, worker_context Context>
-class task {
- public:
-  using value_type = T;
-  using context_type = Context;
-
-  constexpr task(key_t, promise_type<T, Context> *promise) noexcept : m_promise(promise) {}
-
- private:
-  friend constexpr auto get(key_t, task t) noexcept -> promise_type<T, Context> * { return t.m_promise; }
-
-  promise_type<T, Context> *m_promise;
-};
 
 // =============== Final =============== //
 template <worker_context Context>
